requirements.txt
asgiref==3.3.4
attrs==21.2.0
awsebcli==3.19.4
bcrypt==3.2.0
blessed==1.18.0
boto3==1.17.84
botocore==1.20.84
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.7
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-multiple-models==2.1.3
django-storages==1.11.1
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==2.10
importlib-metadata==4.4.0
jmespath==0.10.0
jsonschema==3.2.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
s3transfer==0.4.2
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.59.0
zipp==3.4.1
